Incorporated Linus comments: Split ByzDatabase into separate client and server implementations
